+++ core/index.js
@@ -1,4 +1,5 @@
 module.exports = exports = {
 	terminal: require('./terminal'),
-	sockets: require('./sockets')
+	sockets: require('./sockets'),
+	workers: require('./workers'),
 };